+++ applitools.config.js
@@ -1,0 +1,18 @@
+const { execSync } = require('child_process');
+
+let privateConfig = {};
+try {
+  privateConfig = require('./applitools.private.config.js');
+} catch (e) {}
+
+function getHeadHash() {
+  return execSync('git rev-parse --verify HEAD')
+    .toString()
+    .trim();
+}
+
+module.exports = {
+  ...privateConfig,
+  concurrency: 200,
+  batchId: getHeadHash(),
+};