+++ e2e/tests/constants.js
@@ -6,3 +6,15 @@ export const fixtures = [
   'lists',
   'quote',
 ];
+
+export const DEFAULT_DESKTOP_BROWSERS = [
+  { width: 1440, height: 900, name: 'chrome' },
+  { width: 1440, height: 900, name: 'firefox' },
+  { width: 1440, height: 900, name: 'edge' },
+];
+
+export const DEFAULT_MOBILE_BROWSERS = [
+  { deviceName: 'iPhone X' },
+  { deviceName: 'iPad' },
+  { deviceName: 'Pixel 2' },
+];