+++ e2e/tests/rendering.e2e.js
@@ -1,4 +1,5 @@
 import { fixtures } from './constants';
+import { DEFAULT_DESKTOP_BROWSERS, DEFAULT_MOBILE_BROWSERS } from '../tests/constants';
 
 // function testViewerAndEditorAreEqual() {
 //   function getTitle(test = Cypress.mocha.getRunner().test) {
@@ -20,7 +21,7 @@ describe('editor rendering', () => {
       cy.eyesOpen({
         appName: 'Rendering',
         testName: this.test.parent.title,
-        browser: [{ width: 1440, height: 900, name: 'chrome' }],
+        browser: DEFAULT_DESKTOP_BROWSERS,
       });
     });
 
@@ -36,7 +37,7 @@ describe('editor rendering', () => {
       cy.eyesOpen({
         appName: 'Rendering',
         testName: this.test.parent.title,
-        browser: { deviceName: 'iPhone 6/7/8' },
+        browser: DEFAULT_MOBILE_BROWSERS,
       });
     });