+++ packages/common/web/src/Base/buttons/BlockLinkButton.jsx
@@ -8,7 +8,8 @@ import BlockLinkPanel from './BlockLinkPanel';
 
 class BlockLinkButton extends Component {
   get isActive() {
-    return !!this.props.pubsub.getBlockData({ key: 'componentLink' });
+    const componentData = this.props.pubsub.get('componentData');
+    return !!componentData?.config?.link;
   }
 
   showLinkPanel = () => {