+++ packages/common/web/src/Utils/simplePubsub.js
@@ -25,7 +25,7 @@ export const simplePubsub = initialState => {
     return subscribe(blockHandlerKey(key, blockKey), callback);
   };
 
-  //use to merge objects into store. Merges the the newData with the data for the given key.
+  // Deep merge objects into store. Merges the the newData with the data for the given key.
   const update = (key, newData, blockKey) => {
     const data = get(key);
     const newItem = merge({}, data, newData);